fixed a few bugs w.r.t. non-split histogram part

+++ include/CLsLimit.h
@@ -30,8 +30,8 @@ namespace lands{
 	extern bool _bPositiveSignalStrength;
 	extern vector<double> _lastParams;
 	extern vector<double> _currParams;
-	extern vector< vector< vector<float> > >  vvv_cachPdfValues; // cached pdf values for layers Channel<Process<Event>>>
-	extern vector< float >  v_cachPdfValues; // cached pdf values for layers Channel
+	extern vector< vector< vector<double> > >  vvv_cachPdfValues; // cached pdf values for layers Channel<Process<Event>>>
+	extern vector< double >  v_cachPdfValues; // cached pdf values for layers Channel
 	extern vector< vector< double > >  vv_cachCountingParts; // cached likelihood values for layers Channel<Process>
 	extern vector<  double  >  v_cachHistParts; // cached likelihood values for layers Channel<Process>
 	extern double _countPdfEvaluation;
